rag-backend/requirements.txt
# FastAPI
fastapi
uvicorn[standard]

# Google Gemini
google-genai

# Vector Database
qdrant-client

# Text Processing
langchain-text-splitters

# Configuration
pydantic
pydantic-settings

# Database (as requested)
sqlalchemy
psycopg2-binary
asyncpg
sqlalchemy-utils

# Environment variables
python-dotenv
